Apply explicit type to $result.
HydrateEntity assumes $result will always be some form of object but it's possible for it to be an array also. I've experienced this when implementing the client_credentials grant. [According to the docs](https://laravel.com/api/5.2/Illuminate/Database/Query/Builder.html#method_first) the return type is 'mixed|Builder'.